deps/v8/test/inspector/debugger/async-instrumentation-expected.txt

Checks async instrumentation enabled in the middle.

Running test: beforeAsyncTaskScheduled
test (test.js:16:2)
(anonymous) (expr1.js:0:0)

test (test.js:21:2)
(anonymous) (expr1.js:0:0)

foo (test.js:10:2)
-- Promise.resolve --
test (test.js:19:14)
(anonymous) (expr1.js:0:0)

foo (test.js:12:2)
-- Promise.resolve --
test (test.js:19:14)
(anonymous) (expr1.js:0:0)


Running test: afterAsyncTaskScheduled
test (test.js:16:2)
(anonymous) (expr1.js:0:0)

test (test.js:21:2)
(anonymous) (expr1.js:0:0)

foo (test.js:10:2)

foo (test.js:12:2)


Running test: afterAsyncTaskStarted
test (test.js:16:2)
(anonymous) (expr1.js:0:0)

test (test.js:21:2)
(anonymous) (expr1.js:0:0)

foo (test.js:10:2)

foo (test.js:12:2)
